Remove tm stuff from message processing and place in ticketmaster
