+++ message_processesing.js
@@ -1,5 +1,6 @@
 const request = require('request');
 const PAGE_ACCESS_TOKEN = process.env.PAGE_ACCESS_TOKEN;
+var ticketmaster = require('helpers/ticketmaster.js');
 
 module.exports.handleMessage = function(sender_psid, received_message) {
   if (received_message.text) {
@@ -18,7 +19,7 @@ module.exports.handleMessage = function(sender_psid, received_message) {
     console.log("Location Quick Reply received.");
     console.log(received_message.attachments);
     callSendAPI(sender_psid, {"text": "Finding Events!"});
-    createEventList(sender_psid, received_message);
+    ticketmaster.createEventList(request, sender_psid, received_message);
   } else {
     console.log("Unknown message type, message: " + received_message);
   }
@@ -57,83 +58,4 @@ function callSendAPI(sender_psid, response) {
       console.error("Unable to send message:" + err);
     }
   });
-}
-
-function createEventList(sender_psid, message) {
-  lat = message.attachments[0].payload.coordinates.lat;
-  lng = message.attachments[0].payload.coordinates.long;
-
-  let params = "radius=25&units=miles&latlong="+lat+","+lng+"&apikey="+process.env.TICKETMASTER_APIKEY;
-  let req_url = "https://app.ticketmaster.com/discovery/v2/events.json?"+params;
-  console.log(req_url);
-
-  request({
-      url: req_url,
-      method: "GET"
-    }, (err, res, body) => {
-      if (!err) {
-        var events = JSON.parse(body);
-        if (events["_embedded"] && events["_embedded"]["events"].length > 0) {
-          console.log('ticketmaster requested!');
-          response = generateTMEventTemplate(events["_embedded"]["events"]);
-          console.log(response);
-        } else {
-          response = { "text": "Sorry we couldnt find any events" };
-        }
-        callSendAPI(sender_psid, response);
-      } else {
-        console.error("Unable to send message:" + err);
-      }
-    }
-  );
-}
-
-function generateTMEventTemplate(events) {
-  // Generates a Generic Template for a TicketMaster Event
-  elements = generateElementsJSON(events);
-  return { 
-    "attachment": {
-      "type": "template",
-      "payload": {
-        "template_type": "list",
-        "top_element_style": "compact",
-        "elements": elements,
-        "buttons": [
-          {
-            "type": "web_url",
-            "title": "Refine Search",
-            "url": "https://xandchill.herokuapp.com/refine.html",
-            "webview_height_ratio": "tall",
-            "messenger_extensions": true
-          }
-        ]  
-      }
-    }
-  }
-}
-
-function generateElementsJSON(events){
-  let elements = [],
-      chosenEvents = [];
-
-  for (var i = 0; i < 4; i++) {
-    let randomEventNum = Math.floor(Math.random()*events.length);
-    while(chosenEvents.includes(randomEventNum)){
-      randomEventNum = Math.floor(Math.random()*events.length);
-    }
-    chosenEvents.push(randomEventNum);
-    let event = events[randomEventNum];
-    elements.push({
-      "title": event["name"],
-      "image_url": event["images"][0]["url"],
-      "default_action": {
-        "type": "web_url",
-        "url": event["url"],
-        "messenger_extensions": false,
-        "webview_height_ratio": "compact"
-      }
-    })
-  }
-  console.log(chosenEvents);
-  return elements;
 }
